pyproject.toml
[project]
name = "voronoi-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "cython>=3.1.3",
    "gpjax>=0.12.2",
    "jax[cuda12]>=0.7.1",
    "pillow>=11.3.0",
    "pyright>=1.1.405",
    "rerun-sdk>=0.24.1",
    "scalene>=1.5.54",
    "scikit-learn>=1.7.2",
]
